+++ src/OpenCqrs.EventSourcing/Extensions/ServiceCollectionExtensions.cs
@@ -14,8 +14,8 @@ public static class ServiceCollectionExtensions
     /// </summary>
     /// <param name="services">The service collection.</param>
     /// <param name="types">The types to scan.</param>
-    /// <exception cref="ArgumentNullException">Thrown when services or types is null.</exception>
-    /// <exception cref="InvalidOperationException">Thrown when duplicate type binding keys are encountered.</exception>
+    /// <exception cref="ArgumentNullException">Thrown when services or types are null.</exception>
+    /// <exception cref="InvalidOperationException">Thrown when duplicate type-binding keys are encountered.</exception>
     public static void AddOpenCqrsEventSourcing(this IServiceCollection services, params Type[] types)
     {
         var eventTypeBindings = new Dictionary<string, Type>();