+++ docs/Release-Notes.md
@@ -2,7 +2,7 @@
 
 ## OpenCQRS 7.1.3
 _**Released 13/09/2025**_
-- Fix issue when updating an aggregate that does not exist
+- Update aggregate stores a new aggregate if it doesn't exist
 - Minor improvements
 
 ## OpenCQRS 7.1.2