+++ docs/Release-Notes.md
@@ -1,5 +1,10 @@
 # Release Notes
 
+## OpenCQRS 7.1.3
+_**Released 13/09/2025**_
+- UpdateAggregate stores a new aggregate if it doesn't exist
+- Minor improvements
+
 ## OpenCQRS 7.1.2
 _**Released 10/09/2025**_
 - Rename IDomainEvent to IEvent